Replace parameter SequenceRequest and standalone request parameters with SearchRequest object in SubscriptionEntryCollectionResource::get method

+++ myreader-backend/src/main/java/spring/data/domain/Sequence.java
@@ -15,6 +15,6 @@ public interface Sequence<T> extends Iterable<T> {
 
     boolean hasNext();
 
-    int getPageSize();
+    long getPageSize();
 
 }
